Lots of stuff, made game so much better
Made the first "Level". Adjusted point pingers size and how they activate. Made a smooth camera transition with a stark ball teleportation, seems perfect to me. Added a new font, lightened the mood. Using velocity instead of force for a more dynamic shoot. Added a force minigame where you have to time it correctly. Lives actually work.

Things that need to be worked out are on Trello but off the top of my head Main Menu, endless mode, and a goal to hit in a non-endless mode. Make life counters and possibly a sprite for said lives.

-Ashley (The only person working on this game)