Added Menu and smooth camera transitions
Zooming - When you zoom in and out of the ball it will now use an ienum to be a bit smoother. I zoomed out the zoom in one a bit to make it more useful. Hitting tab will also change perspectives so you can do it easier mid game. Made zoom in faster at beginning.

Modes and Menu - Now a main menu with two options, endless and goal mode. If goal mode is selected the game ends at 10,000 score or if endless is chosen, it will go on forever. Default is endless mode when starting from the second scene.